Move pointer flow graph types into pointer_flow
